+++ src/unescape/mod.rs
@@ -56,7 +56,7 @@ pub fn unescape<'a, S: Into<Cow<'a, str>>>(escaped: S) -> Cow<'a, str> {
 
     #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
     return internal::unescape_in(
-        (internal::Phf, internal::ContextGeneral),
+        (internal::Map, internal::ContextGeneral),
         escaped,
     );
 }
@@ -86,7 +86,7 @@ pub fn unescape_attribute<'a, S: Into<Cow<'a, str>>>(
 
     #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
     return internal::unescape_in(
-        (internal::Phf, internal::ContextAttribute),
+        (internal::Map, internal::ContextAttribute),
         escaped,
     );
 }
@@ -139,7 +139,7 @@ pub fn unescape_in<'a, S: Into<Cow<'a, str>>>(
 
             #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
             return internal::unescape_in(
-                (internal::Phf, internal::ContextAttribute),
+                (internal::Map, internal::ContextAttribute),
                 escaped,
             );
         }
@@ -152,7 +152,7 @@ pub fn unescape_in<'a, S: Into<Cow<'a, str>>>(
 
             #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
             return internal::unescape_in(
-                (internal::Phf, internal::ContextGeneral),
+                (internal::Map, internal::ContextGeneral),
                 escaped,
             );
         }
@@ -207,7 +207,7 @@ pub fn unescape_bytes_in<'a, S: Into<Cow<'a, [u8]>>>(
 
             #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
             return internal::unescape_bytes_in(
-                (internal::Phf, internal::ContextAttribute),
+                (internal::Map, internal::ContextAttribute),
                 escaped,
             );
         }
@@ -220,7 +220,7 @@ pub fn unescape_bytes_in<'a, S: Into<Cow<'a, [u8]>>>(
 
             #[cfg(all(feature = "unescape", not(feature = "unescape_fast")))]
             return internal::unescape_bytes_in(
-                (internal::Phf, internal::ContextGeneral),
+                (internal::Map, internal::ContextGeneral),
                 escaped,
             );
         }